Adding Tx SSB count and measured RSRP value in the script.

See merge request aburger/openairinterface5g!180

+++ ci-scripts/run_sl_usrp.py
@@ -29,7 +29,7 @@ import glob
 import subprocess
 from subprocess import Popen
 import threading
-from typing import Dict
+from typing import Dict, Generator
 from queue import *
 
 HOME_DIR = os.path.expanduser( '~' )
@@ -49,7 +49,7 @@ parser.add_argument('--host', default='10.1.1.80', type=str, help="""
 Nearby Host IP (default: %(default)s)
 """)
 
-parser.add_argument('--user', '-u',  default='', type=str, help="""
+parser.add_argument('--user', '-u',  default='zaid', type=str, help="""
 User id in Nearby Host (default: %(default)s)
 """)
 
@@ -61,16 +61,25 @@ parser.add_argument('--basic', '-b', action='store_true', help="""
 Basic test with basic shell commands
 """)
 
-parser.add_argument('--commands', '-c', default='sl_usrp_cmds.txt', help="""
+parser.add_argument('--commands', default='sl_usrp_cmds.txt', help="""
 The USRP Commands .txt file (default: %(default)s)
 """)
 
-parser.add_argument('--duration', '-d', metavar='SECONDS', type=int, default=25, help="""
+parser.add_argument('--duration', '-d', metavar='SECONDS', type=int, default=30, help="""
 How long to run the test before stopping to examine the logs
 """)
 
 parser.add_argument('--log-dir', default=HOME_DIR, help="""
 Where to store log files
+""")
+
+parser.add_argument('--compress', action='store_true', help="""
+Compress the log files in the --log-dir
+""")
+
+parser.add_argument('--no-run', '-n', action='store_true', help="""
+Don't run the test, only examine the logs in the --log-dir
+directory from a previous run of the test
 """)
 
 parser.add_argument('--debug', action='store_true', help="""
@@ -166,12 +175,12 @@ class TestThread(threading.Thread):
     """
     Represents TestThread
     """
-    def __init__(self, job, queue, commands, passed):
+    def __init__(self, queue, commands, passed):
         threading.Thread.__init__(self)
-        self.name = job
         self.queue = queue
         self.commands = commands
         self.passed = passed
+        self.delay = 0
 
     def run(self):
         if self.queue.empty() == True:
@@ -186,17 +195,16 @@ class TestThread(threading.Thread):
                     thread_delay(job, delay = 0)
                     nearby_proc = self.launch_nearby(job)
                     LOGGER.info(f"nearby_proc = {nearby_proc}")
-                if "syncref" == job:
-                    thread_delay(job, delay = 3)
+                if "syncref" == job and not OPTS.no_run:
+                    thread_delay(job, delay = self.delay)
                     syncref_proc = self.launch_syncref(job)
                     LOGGER.info(f"syncref_proc = {syncref_proc}")
-            if not OPTS.basic:
+            if not OPTS.basic and not OPTS.no_run:
                 LOGGER.info(f"Process running... {job}")
                 time.sleep(OPTS.duration)
                 if nearby_proc:
                     self.kill_process("nearby", nearby_proc)
                 if syncref_proc:
-                    #time.sleep(OPTS.duration)
                     self.kill_process("syncref", syncref_proc)
             self.queue.task_done()
         except Exception as inst:
@@ -213,18 +221,23 @@ class TestThread(threading.Thread):
                     stderr=subprocess.PIPE)
         remote_output = proc.stdout.readlines()
         if remote_output == []:
-            error = proc.stderr.readlines()
-            for raw_line in error:
-                line = raw_line.decode()
-                LOGGER.info(line.strip())
-                if 'PASSED' in line:
-                    self.passed += ['passed']
+            remote_log = proc.stderr.readlines()
         else:
-            for raw_line in remote_output:
-                line = raw_line.decode()
-                LOGGER.info(line.strip())
-                if 'PASSED' in line:
-                    self.passed += ['passed']
+            remote_log = remote_output
+        result_metric = None
+        for raw_line in remote_log:
+            line = raw_line.decode()
+            LOGGER.info(line.strip())
+            # 'SyncRef UE found. RSRP: -100 dBm/RE. It took {delta_time_s} seconds'
+            if 'It took' in line and 'seconds' in line:
+                fields = line.split(maxsplit=12)
+                if len(fields) > 6:
+                    ssb_rsrp = float(fields[-6])
+                    sync_duration = float(fields[-2])
+                    counting_duration = sync_duration - self.delay
+                    result_metric = (ssb_rsrp, sync_duration, counting_duration)
+            if 'PASSED' in line:
+                self.passed += [result_metric]
         return proc
 
     def launch_syncref(self, job) -> Popen:
@@ -249,6 +262,51 @@ class TestThread(threading.Thread):
             proc.wait()
         LOGGER.info(f'kill main simulation processes...done for {job}')
 
+# ----------------------------------------------------------------------------
+
+def get_lines(filename: str) -> Generator[str, None, None]:
+    """
+    Yield each line of the given log file (.bz2 compressed log file if -c flag is used.)
+    """
+    fh = bz2.open(filename, 'rb') if OPTS.compress else open(filename, 'rb')
+    for line_bytes in fh:
+        line = line_bytes.decode('utf-8', 'backslashreplace')
+        line = line.rstrip('\r\n')
+        yield line
+
+def get_analysis_messages(filename: str) -> Generator[str, None, None]:
+    """
+    Finding all logs in the log file with X fields for log parsing optimization
+    """
+    LOGGER.info('Scanning %s', filename)
+    for line in get_lines(filename):
+            #796811.532881 [NR_PHY] nrUE configured
+            #796821.854505 [NR_PHY] PSBCH SL generation started
+            fields = line.split(maxsplit=5)
+            if len(fields) == 4 or len(fields) == 6 :
+                yield line
+
+def analyze_logs(counting_delta: float) -> int:
+    time_start_s, time_end_s = -1, -1
+    log_file = log_file_path
+    sum_ssb = 0
+
+    if OPTS.compress:
+        log_file = f'{log_file_path}.bz2'
+    for line in get_analysis_messages(log_file):
+        #796811.532881 [NR_PHY] nrUE configured
+        #796821.854505 [NR_PHY] PSBCH SL generation started
+        if time_start_s == -1 and 'nrUE configured' in line:
+            fields = line.split(maxsplit=2)
+            time_start_s = float(fields[0])
+            time_end_s = time_start_s + counting_delta
+        if 'PSBCH SL generation started' in line:
+            fields = line.split(maxsplit=2)
+            time_st = float(''.join([ch for ch in fields[0] if ch.isnumeric() or ch =='.']))
+            if time_st < time_end_s:
+                sum_ssb += 1
+    return sum_ssb
+
 def main() -> int:
     commands = Command(OPTS.commands)
     LOGGER.debug(f'Number of iterations {OPTS.repeat}')
@@ -256,26 +314,37 @@ def main() -> int:
         for role, cmd in commands.usrp_cmds.items():
             LOGGER.debug(f'{role} UE: {cmd}')
     jobs = ['nearby', 'syncref'] if OPTS.launch == 'both' else [OPTS.launch]
-    passed = []
+    passed_metric = []
+    num_tx_ssb = []
     num_passed = 0
     for i in range(OPTS.repeat):
         threads = []
         queue = Queue()
         for job in jobs:
             queue.put(job)
-            th = TestThread(job, queue, commands, passed)
+            th = TestThread(queue, commands, passed_metric)
             th.setDaemon(True)
             th.start()
             threads.append(th)
         for th in threads:
             th.join()
-        if num_passed != len(passed):
-            LOGGER.info(f"Passed at the trial {i+1}/{OPTS.repeat}")
+        if num_passed != len(passed_metric):
+            # Examine the logs to determine if the test passed
+            (ssb_rsrp, sync_duration, counting_duration) = passed_metric[-1]
+            num_ssb = analyze_logs(counting_duration)
+            num_tx_ssb += [num_ssb]
+            LOGGER.info(f"Trial {i+1}/{OPTS.repeat} PASSED. {num_ssb} SSB(s) were generated. Measured {ssb_rsrp} RSRP (dbm/RE)")
         else:
             LOGGER.info(f"Failed at the trial {i+1}/{OPTS.repeat}")
-        num_passed = len(passed)
+        num_passed = len(passed_metric)
+
     LOGGER.info('#' * 42)
-    LOGGER.info(f"Number of passed = {len(passed)}/{OPTS.repeat}")
+    LOGGER.info(f"Number of passed = {len(passed_metric)}/{OPTS.repeat}")
+    if len(num_tx_ssb) > 0:
+        LOGGER.info(f"Avg number of SSB = {sum(num_tx_ssb) / len(num_tx_ssb)} ({num_tx_ssb})")
+    if len(passed_metric) > 0:
+        LOGGER.info(f"Avg SSB RSRP = {sum([result[0] for result in passed_metric]) / len(passed_metric)}")
+        LOGGER.info(f"Avg Sync duration (seconds) = {sum([result[1] for result in passed_metric]) / len(passed_metric)}")
     return 0
 
 sys.exit(main())
